+++ src/scripts/create_hellaswag.py
@@ -32,6 +32,7 @@ def main() -> None:
         "en": "Choices",
         "fr": "Choix",
         "it": "Scelte",
+        "es": "Opciones",
     }
 
     for language in choices_mapping.keys():