+++ src/scripts/create_mlsum_de.py
@@ -1,4 +1,4 @@
-"""Create the MLSum-mini summarisation dataset."""
+"""Create the German MLSum-mini summarisation dataset."""
 
 import pandas as pd
 from constants import MAX_NUM_CHARS_IN_ARTICLE, MIN_NUM_CHARS_IN_ARTICLE
@@ -8,7 +8,7 @@ from requests import HTTPError
 
 
 def main() -> None:
-    """Create the MLSum-mini summarisation dataset and upload to HF Hub."""
+    """Create the German MLSum-mini summarisation dataset and upload to HF Hub."""
     dataset_id = "GEM/mlsum"
 
     dataset = load_dataset(dataset_id, "de", token=True)
@@ -60,7 +60,7 @@ def main() -> None:
     )
 
     # Create dataset ID
-    mini_dataset_id = "EuroEval/mlsum-mini"
+    mini_dataset_id = "EuroEval/mlsum-de-mini"
 
     # Remove the dataset from Hugging Face Hub if it already exists
     try: